Improv: Move datatypes to a standalone file
