+++ src/hash.ts
@@ -1,7 +1,8 @@
 import BN from 'bn.js';
 import { sha256 } from 'js-sha256';
 
-import { MPI, concatUint8Array, Byte } from './msgs';
+import { MPI, Byte } from './dataTypes';
+import { concatUint8Array } from './utils';
 
 export function smpHash(version: number, ...args: BN[]): BN {
   let res = new Byte(version).serialize();
